requirements.txt
python-dotenv
crewai
groq
